update: 사이드바 컴포넌트 이름 변경 및 추가 구현
